requirements.txt

numpy>=1.19.0
pillow>=8.0.0
scikit-learn>=0.24.0
opencv-python>=4.5.0
eel>=0.14.0
